+++ src/main/java/com/meongnyangerang/meongnyangerang/controller/NotificationController.java
@@ -2,11 +2,12 @@ package com.meongnyangerang.meongnyangerang.controller;
 
 import com.meongnyangerang.meongnyangerang.dto.notification.MessageNotificationRequest;
 import com.meongnyangerang.meongnyangerang.security.UserDetailsImpl;
-import com.meongnyangerang.meongnyangerang.service.NotificationService;
+import com.meongnyangerang.meongnyangerang.service.notification.NotificationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
